+++ docs/source/lagom.rst
@@ -30,6 +30,3 @@ Evolution Strategies
     
 .. autoclass:: CEM
     :members:
-    
-.. autoclass:: OpenAIES
-    :members:
